Add files via upload
This is the demonstrate a simple DNS labs setup using the cisco packet tracer. this lab is going to show Dns trans late data to to ip addressing. Also the lab is made up of 
(1) server with an ip address (192.168.1.100); (6) PC  with ip address ranging between (192.168.1.100 to 192.168.1.106
and a switch to add or make the connection.